+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update log export example to not use root logger ([#4090](https://github.com/open-telemetry/opentelemetry-python/pull/4090))
 - sdk: Add OS resource detector
   ([#3992](https://github.com/open-telemetry/opentelemetry-python/pull/3992))
+- Update semantic conventions to version 1.27.0
+  ([#4104](https://github.com/open-telemetry/opentelemetry-python/pull/4104))
 
 ## Version 1.26.0/0.47b0 (2024-07-25)